Leak from Intralism Database - Demo Sample
A Sample of a Demo Structure from 5.11.2019, might be outdated. But gives an insight into how easy it actually is to bypass the Process List Gathering. You can edit the Demo out of the Memory. Thought it would be useful to see how one looks like that landed onto the server and is ready to be reviewed by their tool. This one got leaked from their Database.